# PERF: MultiIndex slicing