--- a/asv_bench/benchmarks/indexing.py
+++ b/asv_bench/benchmarks/indexing.py
@@ -211,43 +211,56 @@ class MultiIndexing:
             mi = MultiIndex.from_product([range(1000)] * self.ndim)
         self.df = DataFrame(np.random.randn(len(mi)), index=mi)
 
+        self.tgt_slice = slice(200, 800)
+        self.tgt_null_slice = slice(None)
+        self.tgt_list = list(range(0, 1000, 10))
+        self.tgt_scalar = 500
+
+        bool_indexer = np.zeros(len(mi), dtype=np.bool_)
+        bool_indexer[slice(0, len(mi), 100)] = True
+        self.tgt_bool_indexer = bool_indexer
+
+    def time_loc_partial_key_slice(self):
+        self.df.loc[self.tgt_slice, :]
+
+    def time_loc_partial_key_null_slice(self):
+        self.df.loc[self.tgt_null_slice, :]
+
+    def time_loc_partial_key_list(self):
+        self.df.loc[self.tgt_list, :]
+
+    def time_loc_partial_key_scalar(self):
+        self.df.loc[self.tgt_scalar, :]
+
+    def time_loc_partial_bool_indexer(self):
+        self.df.loc[self.tgt_bool_indexer, :]
+
     def time_loc_all_slices(self):
-        target = slice(200, 800)
-        target = tuple([target] * self.ndim)
+        target = tuple([self.tgt_slice] * self.ndim)
         self.df.loc[target, :]
 
     def time_loc_all_null_slices(self):
-        target = slice(None)
-        target = tuple([target] * self.ndim)
+        target = tuple([self.tgt_null_slice] * self.ndim)
         self.df.loc[target, :]
 
     def time_loc_all_lists(self):
-        target = list(range(0, 1000, 10))
-        target = tuple([target] * self.ndim)
+        target = tuple([self.tgt_list] * self.ndim)
         self.df.loc[target, :]
 
     def time_loc_all_scalars(self):
-        target = tuple([500] * self.ndim)
+        target = tuple([self.tgt_scalar] * self.ndim)
+        self.df.loc[target, :]
+
+    def time_loc_all_bool_indexers(self):
+        target = tuple([self.tgt_bool_indexer] * self.ndim)
         self.df.loc[target, :]
 
     def time_loc_slice_plus_null_slice(self):
-        target = (slice(200, 800), slice(None))
+        target = (self.tgt_slice, self.tgt_null_slice)
         self.df.loc[target, :]
 
-    def time_loc_partial_key_slice(self):
-        target = slice(200, 800)
-        self.df.loc[target, :]
-
-    def time_loc_partial_key_null_slice(self):
-        target = slice(None)
-        self.df.loc[target, :]
-
-    def time_loc_partial_key_list(self):
-        target = list(range(0, 1000, 10))
-        self.df.loc[target, :]
-
-    def time_loc_partial_key_scalar(self):
-        target = 500
+    def time_loc_null_slice_plus_slice(self):
+        target = (self.tgt_null_slice, self.tgt_slice)
         self.df.loc[target, :]