@lukemanley 2022-02-20
I just expanded the asvs for MultiIndex indexing as part of this. There are a lot of code paths for multi indexing and the existing asvs only covered a few of them. A few of the added asvs are currently a bit above the targeted 80ms, but I have a few follow on PRs that will hopefully put them all well below 80ms.